+++ syncapi/storage/sqlite3/output_room_events_table.go
@@ -40,7 +40,7 @@ CREATE TABLE IF NOT EXISTS syncapi_output_room_events (
   -- An incrementing ID which denotes the position in the log that this event resides at.
   -- NB: 'serial' makes no guarantees to increment by 1 every time, only that it increments.
   --     This isn't a problem for us since we just want to order by this field.
-  id BIGINT PRIMARY KEY AUTOINCREMENT, -- DEFAULT nextval('syncapi_stream_id'),
+  id INTEGER PRIMARY KEY AUTOINCREMENT, -- DEFAULT nextval('syncapi_stream_id'),
   -- The event ID for the event
   event_id TEXT NOT NULL UNIQUE, -- CONSTRAINT syncapi_event_id_idx UNIQUE,
   -- The 'room_id' key for the event.
